Jdev

+++ Assets/Scripts/Interfaces/IDmgAndHpInterface.cs
@@ -3,6 +3,6 @@ using System.Collections.Generic;
 using UnityEngine;
 
 public interface IDmgAndHpInterface {
-	public void TakeDamage(int damageAmount);
+	public void TakeDamage(float damageAmount);
 	// public void HealDamage(int healAmount);
 }
